2017/AdventOfCode2017.Test/Day19Test.cs
using Xunit;

namespace AdventOfCode2017.Test
{
    [Trait("Day", "19")]
    public class Day19Test : DayTest<Day19>
    {
        [Theory]
        [InlineData(@"     |          
     |  +--+    
     A  |  C    
 F---|----E|--+ 
     |  |  |  D 
     +B-+  +--+ 
                ", "ABCDEF")]
        public void Part1(string input, string expected)
        {
            Assert.Equal(expected, this.RunPart1(input));
        }

        [Theory]
        [InlineData(@"     |          
     |  +--+    
     A  |  C    
 F---|----E|--+ 
     |  |  |  D 
     +B-+  +--+ 
                ", "38")]
        public void Part2(string input, string expected)
        {
            Assert.Equal(expected, this.RunPart2(input));
        }
    }
}
